# config-control-service/spec/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Config Control Service API",
    "description": "Centralized configuration management and drift detection service.\n\n**Features:**\n- Service instance heartbeat tracking\n- Configuration drift detection and reporting\n- Service discovery integration with Consul\n- Config refresh orchestration via Kafka\n- Real-time monitoring and alerts\n\n**Architecture:**\n- Hexagonal architecture with distinct API, Application, Domain, and Infrastructure layers\n- MongoDB for persistence\n- Redis for caching\n- Kafka for event broadcasting\n",
    "contact": {
      "name": "Platform Team",
      "email": "platform@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "development server"
    }
  ],
  "security": [
    {
      "oauth2_auth_code": [],
      "oauth2_password": []
    }
  ],
...
  "paths": {
...
    "/api/application-services/{id}": {
      "get": {
        "tags": [
          "Application Services"
        ],
        "summary": "Get application service by ID",
        "description": "Retrieve a specific application service by its ID.\n\n**Access Control:**\n- **Authentication required** - Unauthenticated requests will receive 401\n- **Visibility rules apply** - Users can only access:\n  - Orphaned services (ownerTeamId=null)\n  - Services owned by their teams\n  - Services shared to their teams\n- Returns 404 if service doesn't exist or user lacks permission\n",
        "operationId": "findApplicationServiceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application service ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "sample-service"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Application service not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Application service found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationServiceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "put": {
        "tags": [
          "Application Services"
        ],
        "summary": "Update application service",
        "description": "Update an existing application service.\n\n**Required Permissions:**\n- SYS_ADMIN: Can update any service\n- Team members: Can update services owned by their team\n- Partial updates are supported (only provided fields are updated)\n",
        "operationId": "updateApplicationService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application service ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "sample-service"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationServiceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Application service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Application service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationServiceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Application Services"
        ],
        "summary": "Delete application service",
        "description": "Delete an application service permanently.\n\n**Required Permissions:**\n- SYS_ADMIN: Only system administrators can delete services\n- This action is irreversible and will remove all associated data\n",
        "operationId": "deleteApplicationService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application service ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "payment-service"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Application service deleted successfully"
          },
          "404": {
            "description": "Application service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - SYS_ADMIN role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
...
    "/api/application-services": {
      "get": {
        "tags": [
          "Application Services"
        ],
        "summary": "List all application services",
        "description": "Retrieve a paginated list of application services visible to the authenticated user.\n\n**Access Control & Visibility Rules:**\n- **Authentication required** - Unauthenticated requests will receive 401\n- **System admins** see all services (no filtering)\n- **Regular users** see:\n  - Orphaned services (ownerTeamId=null) - for ownership request workflow\n  - Services owned by their teams\n  - Services shared to their teams via ServiceShare grants\n- Server-side filtering ensures users only see authorized services\n\n**Use Cases:**\n- Service discovery and catalog browsing\n- Requesting ownership of orphaned services\n- Viewing team-owned and shared services\n",
        "operationId": "findAllApplicationServices",
        "parameters": [
          {
            "name": "ownerTeamId",
            "in": "query",
            "description": "Filter by owner team ID",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string",
              "description": "Filter by owner team ID",
              "example": "team_core"
            },
            "example": "team_core"
          },
          {
            "name": "lifecycle",
            "in": "query",
            "description": "Filter by lifecycle stage",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string",
              "description": "Filter by lifecycle stage",
              "example": "ACTIVE",
              "enum": [
                "ACTIVE",
                "DEPRECATED",
                "RETIRED"
              ]
            },
            "example": "ACTIVE"
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filter by tags",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by tags",
              "example": [
                "microservice",
                "payment"
              ]
            },
            "example": [
              "microservice",
              "payment"
            ]
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for service name or description",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 0,
              "type": "string",
              "description": "Search term for service name or description",
              "example": "payment"
            },
            "example": "payment"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved application services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationServicePageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "post": {
        "tags": [
          "Application Services"
        ],
        "summary": "Create a new application service",
        "description": "Create a new application service with team ownership.\n\n**Required Permissions:**\n- SYS_ADMIN: Can create services for any team\n- Team members: Can create services for their own team\n- Service ID must be unique across the system\n",
        "operationId": "createApplicationService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationServiceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "409": {
            "description": "Conflict - Service ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Application service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationServiceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type",
            "example": "https://api.example.com/problems/validation-error"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem",
            "example": "Request validation failed"
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/application-services"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp when the error occurred",
            "format": "date-time",
            "example": "2024-01-15T10:30:45.123Z"
          },
          "traceId": {
            "type": "string",
            "description": "Trace ID for correlation across services",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "context": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "Additional context about the error"
            },
            "description": "Additional context about the error"
          },
          "validationErrors": {
            "type": "array",
            "description": "Validation errors (for validation failures)",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "description": "Standardized error response following RFC-7807 Problem Details for HTTP APIs"
      },
...
    },
    "securitySchemes": {
      "oauth2_auth_code": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/auth",
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      },
      "oauth2_password": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  }
}